+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 Added support for Drafter 4.x and 5.x, support for Drafter 3 is retained in
 Draughtsman 0.1.0.
 
+### Enhancements
+
+- Added support for generating source maps using the `generate_source_map`
+  argument to `parse`.
+
 ## 0.1.0
 
 Initial Release.